requirements.txt
flask
flask-cors
flask-sqlalchemy
sqlalchemy-serializer
bcrypt
flask-jwt-extended